backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
scipy==1.13.1
numpy==1.26.4
shapely==2.0.4
geopandas==1.0.1
pyproj==3.6.1
python-multipart==0.0.9
google-generativeai
